requirements.txt
pytorch
flask
tldextract
nltk